+++ frontend/src/App.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Playlists from './components/main/Playlists';
-import JobTrackerFab from './components/JobTrackerFab';
-import JobNotifications from './components/JobNotifications';
+import JobTrackerFab from './components/job/JobTrackerFab';
+import JobNotifications from './components/job/JobNotifications';
 import './App.css';
 
 function App() {
