+++ frontend/src/components/job/JobTrackerFab.tsx
@@ -8,7 +8,7 @@ import {
 import {
   Work as WorkIcon,
 } from '@mui/icons-material';
-import { useJobTracker } from '../lib/useJobTracker';
+import { useJobTracker } from '../../lib/useJobTracker';
 import JobTrackerDrawer from './JobTrackerDrawer';
 
 const JobTrackerFab: React.FC = () => {
